SupprimerAbsence done

+++ client/src/app/site/gestion-absences/gestion-absences.component.ts
@@ -4,6 +4,8 @@ import { CompteurAbsences } from 'src/app/models/compteur-absences';
 import { AbsenceService } from 'src/app/services/absence.service';
 import { StatutAbsence } from 'src/app/models/statut-absence';
 import { typeAbsenceLabels, statutAbsenceLabels } from 'src/app/localisation/french';
+import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
+import { DeleteModalComponent } from './delete-modal/delete-modal.component';
 
 @Component({
   selector: 'app-gestion-absences',
@@ -18,7 +20,7 @@ export class GestionAbsencesComponent implements OnInit
   typeAbsenceLabels = typeAbsenceLabels;
   statutAbsenceLabels = statutAbsenceLabels;
 
-  constructor(private absenceService: AbsenceService) { }
+  constructor(private absenceService: AbsenceService, private modalService: NgbModal) { }
 
   ngOnInit(): void
   {
@@ -35,6 +37,14 @@ export class GestionAbsencesComponent implements OnInit
 
   delete(absence: Absence)
   {
-    this.absenceService.delete(absence).subscribe(() => this.absences.splice(this.absences.indexOf(absence), 1));
+    const modal = this.modalService.open(DeleteModalComponent);
+    const modalComponent = modal.componentInstance as DeleteModalComponent;
+    modalComponent.absence = absence;
+
+    modal.result.then(confirmed =>
+    {
+      if (confirmed)
+        this.absenceService.delete(absence).subscribe(() => this.absences.splice(this.absences.indexOf(absence), 1));
+    });
   }
 }
